docs(stepper): update demo
