--- src/stepper/_example/disable-input.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Stepper } from 'tdesign-mobile-react';
-
-export default function DisableInput() {
-  return <Stepper theme="grey" defaultValue={3} disableInput></Stepper>;
-}
