--- src/stepper/_example/disabled.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import { Stepper } from 'tdesign-mobile-react';
-
-export default function Disabled() {
-  return <Stepper theme="grey" defaultValue={3} disabled></Stepper>;
-}
